requirements.txt
# Core ML/AI libraries
numpy>=1.24.0
scipy>=1.10.0
torch>=2.0.0
pydantic>=2.0.0

# LLM API clients
openai>=1.0.0
anthropic>=0.40.0

# Data analysis
pandas>=2.0.0
statsmodels>=0.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Utilities
typing-extensions>=4.5.0
